fixed some problems with map load factor in class ClientInitializer.

+++ server/src/main/java/com/jukusoft/mmo/gs/frontend/network/ClientInitializer.java
@@ -13,7 +13,7 @@ import java.util.concurrent.atomic.AtomicLong;
 public class ClientInitializer implements CustomClientInitializer {
 
     protected static final int EXPECTED_CONNECTIONS = 100;
-    protected static final int LOAD_FACTOR = 100;
+    protected static final float LOAD_FACTOR = 100;
 
     protected final RegionManager regionManager;
     protected final AtomicLong lastConnID = new AtomicLong(0);
